backend/requirements.txt
Flask==2.1.1
torch==1.10.2
diffusers==0.3.0
Pillow==9.0.1